fixed session bug
