+++ hashview.rb
@@ -1,31 +1,21 @@
 # encoding: utf-8
-#require 'sinatra/base'
 require 'sinatra'
 require 'sinatra/flash'
 require 'haml'
-#require_relative 'routes/init'
-#require_relative 'helpers/init'
-
-#class HashView < Sinatra::Application
-#class HashView < Sinatra::Base
-#  enable :sessions
-
-#  configure :production do
-#    set :haml, { :ugly=>true }
-#    set :clean_trace, true
-#  end
-
-#  configure :development do
-#    # ...
-#  end
-
-#  helpers do
-#    include Rack::Utils
-#    alias_method :h, :escape_html
-#  end
-#end
-
 require_relative 'models/master'
 require_relative 'helpers/init'
 require_relative 'routes/init'
-#require_relative 'helpers/sinatra_ssl'
+
+enable :sessions
+
+before /^(?!\/(login|register|logout))/ do
+  if !validSession?
+    redirect to('/login')
+  else
+    settings = Settings.first
+    if (settings && settings.hcbinpath.nil?) || settings.nil?
+      flash[:warning] = "Annoying alert! You need to define hashcat\'s binary path in settings first. Do so <a href=/settings>HERE</a>"
+    end
+  end
+end
+
