+++ helpers/init.rb
@@ -1,15 +1,9 @@
 # encoding: utf-8
-#require_relative 'email'
-#HashView.helpers SendEmail
-
-#require_relative 'hashimporter'
-#HashView.helpers HashImporter
-
+require_relative 'email'
+require_relative 'hashimporter'
 require_relative 'hc_stdout_parser'
-#HashView.helpers HcStdoutParser
-
 require_relative 'sinatra_ssl'
 require_relative 'sessions'
 require_relative 'status'
 require_relative 'sanitization'
-#HashView.helpers Sinatra
+require_relative 'build_crack_cmd'
