+++ routes/init.rb
@@ -11,3 +11,4 @@ require_relative 'search'
 require_relative 'downloads'
 require_relative 'main'
 require_relative 'login'
+require_relative 'register'
